Fix handing of wide characters / UTF-8
We now expect JSON and YAML input to be UTF-8 encoded, and encode wide
characters to UTF-8 bytes on output for JSON and YAML encoders.  Without
this, if you return a data structure with wide characters, the response
will blow up because HTTP::Message will croak if the content is not
encoded as bytes.

Fixes #88